Revert "split js by functions, not yet bodypix applied"
