+++ firebasePortalSpace/package.json
@@ -1,11 +1,6 @@
 {
-  "name" : "portal-space",
-  "description": "fun video call with your friends! Warp out ! ",
-  "version" : "1.0.0",
-  "type" : "module",
   "dependencies": {
     "firebase-tools": "9.21.0",
     "react-notifications": "^1.7.2"
   }
 }
-
