--- firebasePortalSpace/public/js/signalling.js
@@ -1,11 +1,0 @@
-export const configuration = {
-    iceServers: [
-        {
-            urls: [
-                'stun:stun1.l.google.com:19302',
-                'stun:stun2.l.google.com:19302',
-            ],
-        },
-    ],
-    iceCandidatePoolSize: 10,
-}
