--- firebasePortalSpace/public/js/utils.js
@@ -1,8 +1,0 @@
-export function copyToClipboard(text) {
-    var t = document.createElement("textarea");
-    document.body.appendChild(t);
-    t.value = text;
-    t.select();
-    document.execCommand('copy');
-    document.body.removeChild(t);
-}
