--- firebasePortalSpace/public/js/videoEffect.js
@@ -1,17 +1,0 @@
-// export function loadBodyPix(){
-//     options = {
-//         multiplier: 0.75,
-//         stride: 32,
-//         quantBytes: 4
-//     }
-
-//     bodyPix.load(options)
-//         .then(net => performance(net))
-//         .catch(err => console.log(err))
-// }
-
-// export async function perform(net){
-//     while(1){
-        
-//     }
-// }
